Address

+++ java/runtime/src/main/java/org/ray/runtime/gcs/RedisClient.java
@@ -104,9 +104,9 @@ public class RedisClient {
     }
   }
 
-  public Long incr(byte[] key) {
+  public int incr(byte[] key) {
     try (Jedis jedis = jedisPool.getResource()) {
-      return jedis.incr(key);
+      return jedis.incr(key).intValue();
     }
   }
 }
